optimizing:IamClientSessionManager validating batch

+++ super-devops-iam/super-devops-iam-client/src/main/java/com/wl4g/devops/iam/client/session/mgt/IamClientSessionManager.java
@@ -21,12 +21,15 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.locks.Lock;
 
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 
 import org.apache.shiro.session.Session;
+import org.springframework.beans.factory.annotation.Autowired;
 
+import static com.wl4g.devops.common.constants.IAMDevOpsConstants.LOCK_SESSION_VALIDATING;
 import static com.wl4g.devops.common.constants.IAMDevOpsConstants.CACHE_TICKET_C;
 import static com.wl4g.devops.iam.client.filter.AbstractClientIamAuthenticationFilter.SAVE_GRANT_TICKET;
 
@@ -36,6 +39,7 @@ import com.wl4g.devops.iam.common.authc.model.SessionValidityAssertModel;
 import com.wl4g.devops.iam.common.cache.IamCacheManager;
 import com.wl4g.devops.iam.common.session.IamSession;
 import com.wl4g.devops.iam.common.session.mgt.AbstractIamSessionManager;
+import com.wl4g.devops.support.concurrent.locks.JedisLockManager;
 import com.wl4g.devops.support.redis.ScanCursor;
 
 /**
@@ -49,14 +53,12 @@ import com.wl4g.devops.support.redis.ScanCursor;
 public class IamClientSessionManager extends AbstractIamSessionManager<IamClientProperties> {
 
 	/**
-	 * Iteration batch size.
-	 */
-	final private static int DEFAULT_BATCH_SIZE = 100;
-
-	/**
 	 * Expire session validator
 	 */
 	final protected IamValidator<SessionValidityAssertModel, SessionValidityAssertModel> validator;
+
+	@Autowired
+	protected JedisLockManager lockManager;
 
 	public IamClientSessionManager(IamClientProperties config, IamCacheManager cacheManager,
 			IamValidator<SessionValidityAssertModel, SessionValidityAssertModel> validator) {
@@ -71,41 +73,58 @@ public class IamClientSessionManager extends AbstractIamSessionManager<IamClient
 
 	@Override
 	public void validateSessions() {
-		log.info("Validating all active sessions...");
+		Lock lock = lockManager.getLock(LOCK_SESSION_VALIDATING);
 
 		try {
-			ScanCursor<IamSession> cursor = sessionDAO.getAccessSessions(DEFAULT_BATCH_SIZE);
-			while (cursor.hasNext()) {
-				List<IamSession> activeSessions = cursor.readValues();
+			if (lock.tryLock()) {
+				log.info("Validating all active sessions...");
 
-				// Grant ticket of local sessions.
-				Map<String, Session> localSessions = new HashMap<>(activeSessions.size());
+				ScanCursor<IamSession> cursor = sessionDAO.getAccessSessions(DEFAULT_SCAN_BATCH_SIZE);
+				while (cursor.hasNext()) {
+					List<IamSession> activeSessions = cursor.readValues();
 
-				// Wrap to validation assertion
-				SessionValidityAssertModel request = new SessionValidityAssertModel(config.getServiceName());
-				for (IamSession session : activeSessions) {
-					String grantTicket = (String) session.getAttribute(SAVE_GRANT_TICKET);
-					request.getTickets().add(grantTicket);
-					localSessions.put(grantTicket, session);
-				}
+					// Grant ticket of local sessions.
+					Map<String, Session> clientSessions = new HashMap<>(activeSessions.size());
 
-				// Validation expires sessions.
-				SessionValidityAssertModel assertion = validator.validate(request);
-				for (String deadTicket : assertion.getTickets()) {
-					Session session = localSessions.get(deadTicket);
-					try {
-						if (nonNull(session)) {
-							sessionDAO.delete(session);
-							log.info("Cleauping expired sessionId: {}", session.getId());
-						}
-					} catch (Exception e) {
-						log.warn("Failed to cleaup expired sessions. sessionId: {}, deadTicket: {}", session.getId(), deadTicket);
+					// Make to validation request
+					SessionValidityAssertModel request = new SessionValidityAssertModel(config.getServiceName());
+					for (IamSession session : activeSessions) {
+						String grantTicket = (String) session.getAttribute(SAVE_GRANT_TICKET);
+						request.getTickets().add(grantTicket);
+						clientSessions.put(grantTicket, session);
 					}
+
+					// Validation expires sessions.
+					SessionValidityAssertModel assertion = validator.validate(request);
+					for (String deadTicket : assertion.getTickets()) {
+						Session session = clientSessions.get(deadTicket);
+						try {
+							if (nonNull(session)) {
+								sessionDAO.delete(session);
+								log.info("Cleauping expired sessionId: {}", session.getId());
+							}
+						} catch (Exception e) {
+							log.warn("Failed to cleaup expired sessions. sessionId: {}, deadTicket: {}", session.getId(),
+									deadTicket);
+						}
+					}
+
 				}
+			} else {
+				log.info("Skip validating all active sessions.");
 			}
+
 		} catch (Exception e) {
 			log.error("Validating expire sessions failed", e);
+		} finally {
+			lock.unlock();
 		}
+
 	}
+
+	/**
+	 * Validating scan iteration batch size.
+	 */
+	final private static int DEFAULT_SCAN_BATCH_SIZE = 1000;
 
 }
